# address-api/internal/address/dto/request/address_patch_request.go
package request

type AddressPatchRequest struct {
	Doc []PatchRequest `json:"doc"`
}

type PatchRequest struct {
	Op    string      `json:"op"`
	Path  string      `json:"path"`
	Value interface{} `json:"value"`
}