# address-api/internal/address/dto/response/address_response.go
package response

type AddressResponse struct {
	Id          int    `json:"id"`
	City        string `json:"city"`
	Country     string `json:"country"`
	FullAddress string `json:"fullAddress"`
	UserId      string `json:"userId"`
}